requirements.txt
requests
beautifulsoup4
selenium
seleniumbase
python-dotenv
